[cipd] Introduce the notion of root prefix ("") and root metadata.
This allows to stop special-casing handling of "administrators" group and of
RPCs that touch root (as in the future ListPrefix("") RPC).

Now "administrators" are just set as OWNERS on the package repository root and
all necessary permissions are just naturally follow from there.

In the future (once PrefixMetadata proto has more metadata knobs), this would
allow to set default settings for all packages.

R=nodir@chromium.org, tandrii@chromium.org
BUG=801423

Change-Id: Id0bd3a4311308042ab685a8d40bc798c5b23add1