[logdog] NACK messages if they are transient failures
Problem: Newly deployed instances of logdog Archivist processes a bunch of transient
failures, and the stops working.
Why? Because the Google Cloud Go library was updated.

Since Google Cloud Pubsub Go Library v0.19.0, the behavior of the subscription
receiver flow control went from:
* Release the flow controller when the function ends.
To:
* Release the flow controller when Ack or Nack is called.

See:
https://github.com/GoogleCloudPlatform/google-cloud-go/issues/919#issuecomment-372403175

Because LogDog doesn't NACK transient failures, this means the flow controller just
sees a bunch of outstanding messages, and decides to wait until a message gets
ACK'ed or NACK'ed.  This never happens.

The other way to fix this is to set MaxOutstanding to -1, but we want to NACK anyways.

Bug: 795156
Change-Id: Ibcc1466494bc1d22e4c7cd3fdaaaa2fcf3902008