[luci-notify] Use gitiles.PagingLog when retrieving commits
This change updates the gitiles history func implementation to use
gitiles.PagingLog instead of just Log when retrieving commits. If we
don't use PagingLog, there's a much higher chance that in the list of
commits we won't see an old-but-in-order commit, which will then be
registered as out-of-order. While this still has a limit on the number
of commits, it's 1000, which is much more reasonable. We should,
however, consider the case where a commit is older than that and reset
the Builder in the datastore appropriately.