Update patch set 2
Patch Set 2:

This was a bug Petr discovered. It should solve the problem for most cases, but since this still relies on the default limit of 1000, it's not perfect and we should probably reset the builder in the datastore if the commit is too old (that is, the RPC succeeds, has hit the well-known max limit of commits, but doesn't actually contain the ancestor revision).

Patch-set: 2
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>