[cloudlogger] Add Stackdriver logging as a logger
This is intended to enable http Request style logging in GKE apps.
the Stackdriver API must be used because the GKE fluentd config does
not allow custom log_names to be set, which is a requirement for combined
logs.

The Stackdriver logging go library is still in alpha, so this
is intended to be used only as a secondary logging stream alongside
logging/teelogger.

Change-Id: I51b0efaebdd2fad23a26b55b54b9f9391c9f637b