[luci-notify] Refactor builder updates
When a builder is about to get updated when handling a pubsub build, it
currently gets updated right before being put. This change instead
inverts that, and creates a copy of the builder when available, updating
the copy with whatever new information we want to store as it becomes
available.