[luci-notify] Make Notify stateful
This change refactors the Notify function which currently has a lot of
state passed in into depending on a stateful struct called Notifier. The
Notifier is intended to be constructed in a builder-pattern-like manner.

Also, unwrap Build as we can now use buildbucketpb.Build again directly,
and derive per-build email notifications from build.input.properties.

Change-Id: Id7d3741bfb08e56d93e31f2ee0e2712e512d19d9