[luci_notify] Allow explicit repository for builder in config
This change allows one to set an explicit repository for a builder in
the luci-notify configuration. The new semantics are as follows:

If a repository is set, then luci-notify will continue to work as
before, except it will ignore all builds which do not have a gitiles
commit from a repository which matches the repository in the
configuration.

If a repository is not set, then luci-notify will use only build
creation time to determine what the "previous" build was.

This change also refactors the code such that Builders in the datastore
are managed by the configuration codepaths rather than by the PubSub
codepaths (now, they only update the Builders and should never create
them).

Finally, this change removes the notion of a Notifier. Although
Notifiers will continue to exist in the project configuration, they will
no longer be used internally by luci-notify. After the restriction that
a Builder could only be listed once per configuration, Builders became
tied to a single Notifier, or a single set of Notifications. Thus,
we opt to place the Notifications directly as part of the Builder.

Change-Id: Ibb1580035d79220a51761adf506f87092479f299
Reviewed-on: https://chromium-review.googlesource.com/1080271
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>