[luci-notify] Refactor Notify into two functions
This change refactors the Notify function which currently has a lot of
state being passed in into ExtractRecipients and Notify, which are then
used in conjunction. It also refactors Notify to accept the email template
input directly, in order to help make the interface more explicitly focused
around notifying, and improve ergonomics slightly.

Change-Id: Id7d3741bfb08e56d93e31f2ee0e2712e512d19d9