[luci-notify] Refactor builder updates
When a builder is about to get updated when handling a pubsub build, it
currently gets updated right before being put. This change instead
inverts that, and creates a copy of the builder when available, updating
the copy with whatever new information we want to store as it becomes
available.

Change-Id: I2ac2ae24a52a15a1bc62bdad4503281bf41d4c51
Reviewed-on: https://chromium-review.googlesource.com/1092102
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>