[milo] Change URLs: /b/:project/b/b:id -> /b/:id
Users need a simple way to translate a buildbucket ID into a human-consumable
build URL. Here "simple" implies without RPCs.

The "project" parameter in "/b/:project/b/b:id" URL path template is dummy
(id is unique anyway, we don't even check that project is correct) and confusing
to users (which project a given build belongs to?). Remove it by changing the
template to "/b/:id". Remove "b" prefix because it turned out to be useless.

Bug: 841603
Change-Id: Icab6914d17aca301c25fe653b3d355ad2de5d8c0