Update patch set 35
Patch Set 35:

(5 comments)

Alrighty, I ended up tearing a lot of this apart and rewriting it.

New semantics are that you can have a blamelist notification config set on as many Notifications as you want, but that you may end up sending duplicates. This required a restructuring of a bunch of things.

Also, with repository whitelists, it's actually possible to customize the template based on the repo you're notifying about now, which is kind of cool. IDK if there's any use-case for it, but these semantics seemed to me to be the most sane of alternatives. Limiting to just one blamelist notification config per notifier seemed overly restrictive, and it's a little awkward in the code itself. This could be written somewhat more cleanly (?).

Anyway, also wrote tests, all of them seem to pass.

Patch-set: 35
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>