[milo] increase masterExpiry to 4h
A master or its network may be offline for an hour. This leads milo to think
that master was deleted, it responds to buildbot API calls as "there are no
builders on this master" which leads other systems (e.g. luci-migration) to make
wrong decisions.

Increase expiration to 4h.

Bug: 852993
Change-Id: I3f7ea76b652e2de667aff54fb828a96643657714
Reviewed-on: https://chromium-review.googlesource.com/1102191
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>