[milo] Fix interaction of linkifying rules with each other and escaping.
formatCommitDesc ran regexes on the HTML escaped input. This results in
a number of incorrect conversions. E.g., rewrote:

    https://crbug.com/618365

to:

    <a href="https://<a href="https://crbug.com/618365">crbug.com/618365</a>">https://<a href="https://crbug.com/618365">crbug.com/618365</a></a>

because it would first apply the URL-linking rule on the full URL, then
it would find several instances of crbug.com/618365, which it dutifully
linked as well. This particular case was even enshrined in a test
expectation.

For a truly entertaining case, see
https://ci.chromium.org/p/chromium/builders/luci.chromium.ci/GPU%20FYI%20Linux%20Builder/5847.
Here it rewrites:

    Bug: https://crbug.com/828864

as:

    Bug: <a href="https://<a href="https://crbug.com/<a href="https://crbug.com/828864">828864</a>">crbug.com/<a href="https://crbug.com/828864">828864</a></a>">https://<a href="https://crbug.com/<a href="https://crbug.com/828864">828864</a>">crbug.com/<a href="https://crbug.com/828864">828864</a></a></a>

There are seven hrefs in there. :-) First, it sees
https://crbug.com/828864 and linkifies that (1). That results in two
crbug.com/828864 instances, which matches the next rule (2). That
results in four 828864 instances, which matches the "Bug: ..." rule,
each of which gets linkified (4). 1 + 2 + 4 = 7.

Instead, use a slightly richer intermediate representation that prevents
already-processed chunks of text from being processed a second time.

Bug: 855719
Change-Id: Ica1aec4ed10ca2547680fca8639a996cd0d529f2
Reviewed-on: https://chromium-review.googlesource.com/1112575
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>