[buildbucket] Add Build.Output.gitiles_commit
Add a field that specifies the commit that was actually checked out and executed
against. This supersedes got_revision property. When GitilesCommit grows
"position" field, it will supersede got_cp as well.

This is different from go/buildbucket-blame: does not include deps.
Using a map of gitiles commits makes it difficult to figure out which of
output commits corresponds to the input commit, and overall which of output
commits to use for blamelist computation. It could be mitigated with
something like Build.Output.main_checkout field, but it brings its own
complexity: different builds are allowed to specify different values,
it is impossible to specify the main checkout in a field mask, so that
RPC response includes only the main checkout. Instead we start with simpler
singular commit. When/if we need to expose DEPS, later we can add
gitiles_commit_deps.

Allow filtering by output commit id or output commit ref.
Filtering by ref is important for builders that want to be able to build
for a arbitrary refs. For such builds, when computing blamelist, the previous
build should be determined with ref filter.

There is no known use case for filtering on input commit.

Change-Id: I0da37cc8ed38bfa9dc86beb1caac67db88127f66