[buildbucket] Add BuildPredicate.build: BuildRange
Add a build range to BuildPredicate. Here range is defined on the order in
context, which is usually newest-to-oldest, but can be extended to order by
commit position. The new field works for both and can be used to find the
"previous" build for the purpose of computing blamelist.

Bug: 851585
Change-Id: Ifbfef84a5cf8711e940111bc4d0ac0849d87c8b9
Reviewed-on: https://chromium-review.googlesource.com/1120433
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>