Only require regexp to start with ref prefix, not literal prefix
The requirement to start with ref prefix was originally added to make it
possible to extract regexp for the suffix. However, we only remove the ref
prefix (literal prefix trimmmed to the last slash), therefore there is no need
to require regexp to start with its entire literal prefix.

Example config that does not start with literal prefix, but can still be used
for matching refs can be found here: https://crrev.com/c/1124455.

R=tandrii@chromium.org

Bug: 836232
Change-Id: I20cd6bdeb2e26ceb6ca14d5097ccb9a4f015e65b