[cipd] Rename ExtractInstance to ExtractFiles, make it accept []File.
When repairing instances, we have []File, but may not have the PackageInstance
(if we didn't fetch anything). But ExtractInstance doesn't really need
PackageInstance. It only uses its Files() method to get []File.

This change necessitates moving .cipd/* gut filtering to Deployer, which is
probably a better place for it anyway.

R=iannucci@chromium.org
CC=jbudorick@chromium.org
BUG=794764

Change-Id: I7fa9dfbe11a809bee1073c576c322826b50ea560