[swarming] Add -tasks-file flag to trigger
This change adds a tasks file flag to the trigger subcommand which
allows passing a file containing a list of Swarming task requests
encoded as JSON to trigger, and triggering multiple tasks at once.

If the flag is specified, all other flags are ignored, as the Swarming
task requests inside of the file are considered complete.

This change also cleans up some of the trigger code with respect to Go
style.

Change-Id: I79fce91ac4e546d83b5ffbe227c37301e64bd34f