Update patch set 9
Patch Set 9:

Sorry for the sloppy code! I rewrote it in the current form in a bit of a hurry, and I haven't had a moment to try and run it end-to-end. I'll do that now.

I also need to (and want to) add some semblance of tests.

Patch-set: 9
Reviewer: Gerrit User 1154295 <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>