[cipd] Support SHA256 for client bootstrap (server side).
The backend recognizes infra/tools/cipd/* packages as ones containing the CIPD
client binary. It extract this binary and stores it in CAS.

The hash algo used to name the extracted file in CAS is made identical to the
algo used to name the original CIPD package file. That way we can be sure that
once we switch to SHA256 hash for the CIPD client package, we'll also switch
to SHA256 for the bootstrap.

For legacy bootstrap scripts and clients, we also always provide SHA1 hash of
the client binary, since that's what is hardcoded in the bootstrap scripts
committed in repositories/embedded in the images, etc. (All this places will
eventually be updated to verify SHA256).

Also, the client itself in its self-update procedure currently checks only SHA1
hash. This will be fixed soon.

R=iannucci@chromium.org, nodir@chromium.org
BUG=821194

Change-Id: I34dbe4f4c4560a79c9e4866b2c38a02321a9733b