[cipd] Calculate and return all supported hashes when extracting the client.
Server side only.

This would allow old clients to self-update when we add new hash algos (if it
ever happens again). For the verification the client would just pick the best
algo it understands and ignore the rest.

Keep LegacySha1 field too, since the clients that use it are already deployed
and can't be undeployed.

This CL also switches server tests for client bootstrap endpoints to use
SHA256.

R=iannucci@chromium.org, nodir@chromium.org
BUG=821194

Change-Id: Ifaf7b7cdb4e4dd4b29ead36b3ed222b34ae44bff