[cipd] Make CIPD client NOT freak out when it sees unknown hash algos.
It now can transparently round trip ObjectRef and instance IDs that use hash
algos not known to the client. This is important in the self-update scenario,
which consists of two RPCs:
  1) ResolveVersion(version, ...) => ObjectRef.
  2) DescribeClient(ObjectRef, ...).

If 'version' happens to point to a package instance that uses some new algo,
the client used to freak out and fail to self-update (even though it can safely
do so, using some hash it does know from client_ref_aliases in DescribeClient
output).

Now the client worries about a hash algo only if it is going to use it for the
hash verification.

R=nodir@chromium.org, iannucci@chromium.org
BUG=821194

Change-Id: I28b8570bab0f6d8870d71700c9be1b173c16ca8b
Reviewed-on: https://chromium-review.googlesource.com/1149614
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>