[cipd] Fix regression in 'cipd search' subcommand.
It used to return 0 exit code and no results if asked to search inside missing
package. With v2 switch it started to return exit code 1 instead, which broke
some callers.

This change restores the original behavior. It sort of makes sense, since if
a package is missing, there's no instances of it, and search should just return
0 results.

R=nodir@chromium.org, iannucci@chromium.org
CC=​phosek@chromium.org
BUG=867729, 801423

Change-Id: I274b35c82be4d11db78150b444b076ef7d9787a0