[cipd] Implement generation and reading of *.digests file.
It is involved during self-update procedure. It holds the expected hashes of the
client binary (per platform). It is supposed to be committed side by side with
scripts that do CIPD client bootstrap, so that they can check the hash of the
file they download.

Three new flavors of 'selfupdate' call:
  * 'cipd selfupdate -version-file <path>': reads the version to update to from
    the given file, and reads expected client digests from <path>.digests, then
    fetches the new client and verifies the hash is know (this last part is
    still TODO).
  * 'cipd selfupdate -version-file <path> -generate-digests': resolves the
    client version, fetches digests of the client binary and writes them into
    <path>.digests.
  * 'cipd selfupdate -version-file <path> -check-digests': returns 0 iff the
    digests file is up-to-date (i.e. regenerating it now is noop).

R=iannucci@chromium.org
BUG=870166

Change-Id: Id73b2326f88562e5e0b91088bde91293fd745148