[cipd] Move *.digests generation and checks into selfupdate-roll subcommand.
Also teach it to also write to the version file, so to roll a pinned client all
we have to do it run:

  cipd selfupdate-roll -version-file <path> -version <new-version>

The primary reason we do it at once is that in depot_tools (and probably other
places where the client is pinned), modifying <version-file> manually makes
'cipd' invocation try to self-update to this new version next time the client
runs. So when running 'cipd selfupdate ... -generate-digests' to regenerate
*.digests we hit the self-update code path with OLD *.digests file, and the
self-update rightfully fails.

R=nodir@chromium.org, iannucci@chromium.org
BUG=870166

Change-Id: I18a57a12fd6b34d45f62ea52036d31575d36f091