README.md: trim stale instructions
The official way to work in luci-go is via an infra checkout, so
document it as this.

Remove appengine/README.md, it was counter-productive.

There's 85 README.md in the repository. I assume they are stale to
various degree and should be replaced with Go documentation.

Change-Id: I8f3366fbd4d379a81c204d7e5b5e7076fab6344d
Reviewed-on: https://chromium-review.googlesource.com/1176128
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>