Update patch set 3
Patch Set 3:

this is because proto files were symlinked and go files are generated from them.
if we are adding a new flag to cproto, ideally we don't have to symlink in the first place.

Besides, the current approach prohibits importing proto files that live in currentGoPath, but under the specified path.

I ran

  protoc --proto_path=/Users/nodir/infra/infra/appengine/monorail --go_out=. /Users/nodir/infra/infra/appengine/monorail/api/api_proto/*.proto

and it generated https://chromium-review.googlesource.com/c/infra/infra/+/1179972

which means it is sufficient to just append the specified path(s) to --proto_path, without affecting existing logic.

If we decided to do so, let's accept a list of extra proto paths, but just one:
https://godoc.org/github.com/luci/luci-go/common/flag/stringlistflag

Patch-set: 3
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>