Update patch set 1
Patch Set 1:

In most/all of these, the long description is going from a single line to multiple lines when output. Instead of inserting \n everywhere, I find it useful to break this out into an easier to read raw string. Example:

const getDatacentersLongDesc = `Retrieves datacenters matching the given names, or all datacenters if names are omitted.

Example to get all datacenters:
  crimson get-dcs

Example to get description of LAB0 datacenter:
  crimson get-dcs -name LAB0
`

And then in the command struct, just use the const:

LongDesc: getDatacentersLongDesc,

At the moment, these are not terrible, but later if more text and/or examples are added, the current style of inserting \n everywhere gets difficult to visualize when reading the code. Modifying the raw string is much easier for anyone to do.

Patch-set: 1
Reviewer: Gerrit User 1300298 <1300298@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>