[buildbucket] Add Step.Log.url
Users need to fetch step log contents using LogDog API. Currently Buildbucket
does not expose logdog address, so some users have to resort to separately
fetching original annotation proto which contains logdog stream addresses.
This info should be available at the buildbucket level.

Add Step.Log.url field that is a URL of the log, as opposed to the URL of the
log viewer. This is intentionally less structured (a string, not a proto).
- minimal design / API surface
+ agnostic of log storage
+ allows both absolutely URLs (with logdog project and prefix), and relative
  (relative to Build.Infra.logdog.*). We will use absolute form anyway, though,
  because it is simpler
- requires more parsing. It should not be a big deal because the field will be
  used only by programs (not humans) that need to fetch log contents. They will
  use a logdog client anyway, which typically has a parser. Otherwise, parsing
  this string isn't hard, programmatically.
  It is difficult in SQL/BQ, but such use case is not anticipated.

Rejected design: stream name field that is supposed to be used in conjunction
with Build.Infra.logdog.{host, project, prefix} fields. Mostly because in
go/buildbucket-api-v2 iannucci@ suggested that in the future we might want to
support lightweight steps which happen on different machines, with different
logdog prefixes. The design proposed in this CL does not have this problem.

Change-Id: Ieb33edfac20edac0a5d9060ba6592ce9a7b40980