[buildbucket] Add Step.Log.url
Users need to fetch step log contents using LogDog API. Currently Buildbucket
does not expose logdog address, so some users have to resort to separately
fetching original annotation proto which contains logdog stream addresses.
This info should be available at the buildbucket level.

Add Step.Log.url field that is a URL of the log, as opposed to the URL of the
log viewer. This is intentionally less structured (a string, not a proto).
+ minimal design / API surface
+ agnostic of log storage
+ allows both absolute URLs (with logdog project and prefix), and relative
  (relative to Build.Infra.logdog.*). We will expose absolute form anyway, though,
  because it is simpler. Impl may decide to store relative URLs if they are relative
  to the common prefix.
- requires more parsing. It should not be a big deal because the field will be
  consumed only by programs (not humans) that need to fetch log contents. They will
  use a logdog client anyway, which typically has a parser. Otherwise, parsing
  this string isn't hard.
  It is hard in SQL/BQ, but such use case is not anticipated.

Rejected design: stream name field that is supposed to be used in conjunction
with Build.Infra.logdog.{host, project, prefix} fields. Mostly because in
go/buildbucket-api-v2 iannucci@ suggested that in the future we might want to
support lightweight steps which happen on different machines, with different
logdog prefixes. The design proposed in this CL does not have this problem.

Bug: 879242
Change-Id: Ieb33edfac20edac0a5d9060ba6592ce9a7b40980
Reviewed-on: https://chromium-review.googlesource.com/1211002
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>