[cipd] Use SHA256 hash algo by default.
It has been used for various packages for about a month now, so it should
probably be fine to use it for all packages.

Later the server will be configured to reject SHA1 uploads. For now it is
still allowed, to be able to quickly deal with possible breakages caused by
SHA256 packages.

BUG=821194
R=nodir@chromium.org

Change-Id: I55d214f19b283b6b044f4c1e1c04861d2b1bd30f