[cloudkms] Properly base64 encode/decode the ciphertext
Previously, we avoided base64 encoding/decoding the ciphertext under the
assumption that this tool would be used on it's own. However, some
ciphertexts are encrypted with the gcloud command which does do this
properly. It was likely an unnecessary assumption that would cause bugs,
so this fixes that.

R=vadimsh@chromium.org

Change-Id: Iab40d2bc3cd6374341c6141fee228041222f4601
Reviewed-on: https://chromium-review.googlesource.com/1227570
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Michael Knyszek <mknyszek@google.com>