[cipd] Handle stale *.digests file in self-update more gracefully.
Before this CL self-update unconditionally trusted what's in *.digests file.
This led to very confusing behavior when manually rolling the client version,
but forgetting to update *.digests file:

1. Bootstrap a client.
2. Manually modify cipd_client_version (skipping 'cipd selfupdate-roll').
3. Run the client self-update. Since the client binary hash still matches what's
   in *.digests, it assumes it is up-to-date. And it happily records its version
   (in .versions/.cipd_client.cipd_version) to match new version from
   cipd_client_version file.
4. The client is left in a semi-broken state: it claims it is new, while in fact it
   is old.

To fix this, self-update now always double checks that the hash in the *.digests
matches what is known to the backend. The mismatch indicates that either the
*.digests file is stale (most likely), or the backend is lying (very unlikely).

Due to caching of client refs in the tag cache, this causes 0 new RPCs on a
happy path (when the client is up-to-date). On the unhappy path,
DescribeClient RPC now happens slightly earlier.

R=nodir@chromium.org, iannucci@chromium.org, tandrii@chromium.org
BUG=870166

Change-Id: I43c614a8e231bcdabed4d6012569b50231105944
Reviewed-on: https://chromium-review.googlesource.com/1226169
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>