[token-server] Fix flake in rpc_inspect_machine_token_test.go.
The test used to replace few bits on a token with zeros to make the signature
"broken".

Some time ago Go introduced randomization to the RSA private key generator
(used in the test) that can't be disabled. Thus the signature also become
random. And sometimes the bits being replaced by zeroes were actually already
zeros.

Replace much more bits with zeros. The chance of a flake should be negligible
now.

R=tandrii@chromium.org
BUG=894676

Change-Id: I35e6326413880eeaf0276951684ee3143ad7565a