[mapper] Refactor how mappers are defined.
We now register a factory function that produces an instance of Mapper before
processing a set of pages. This gives opportunity to prepare state needed by
a mapper only once per mapping task (instead of doing it before each page).

This also makes mappers aware of a job and a shard they are working on. This may
be useful if they want to e.g. submit a named TQ task with name derived from
job/shard IDs.

R=iannucci@chromium.org

Change-Id: Ie130e46e9f57eafb2225ef7377f9a7dc07881eab