[cipd] Add a mapper that discovers broken tags.
Since it's not clear how many broken tags there are in the datastore, or how
costly it is to map over all existing 6M tags, prepare a mapper that just
marks broken tags. We can deal with marked tags later in an appropriate way
(depending on how many broken tags we have).

R=iannucci@chromium.org, nodir@chromium.org
BUG=875064

Change-Id: Id7c59f97bf36857e6a2f1d795280e4d56b8fc29e