[logdog] Remove topic existance check.
Presumably if the topic doesn't exist we'll find out soon enough when
we attempt to publish to it. We've observed these existance checks
flakily failing in the wild. For an existing logdog deployment they
only introduce another potential point of failure, and for new logdog
deployments it's pretty easy to diagnose that the topic is missing and
create it.

R=hinoka@chromium.org, vadimsh@chromium.org

Change-Id: Id74bd42dc9c9853ad5195e526582ab94d0056f4a