[cipd] Use SHA256 to verify integrity of tag and instance cache files.
Once deployed, existing files will be treated as "corrupted" (though with
friendlier logging message).

For the tag cache, it means total cache flush. It isn't a big deal, since tag
resolutions are simple fast RPCs.

For the instance cache, it will trigger a rebuild of the cache index. Cached
instances will NOT be lost.

R=nodir@chromium.org, tandrii@chromium.org, iannucci@chromium.org

Change-Id: I18c553091880b7f93d0d8bdd965e3da7716e32e2