[logdog] Do not schedule optimistic archival
Previously, the behiavor of logdog was to schedule (up to) 2 archival tasks.
For example if ArchiveDelayMax was set to 48hr, and SettleDelay set to 5min:
RegisterPrefix: Schedule archival in 48hrs (pessimistic archival)
TerminateStream: Schedule archival in 5 min, replace pessimistic archival (optimistic archival)

This CL removes the optimistic archival, so all stream are archived 48hrs
after registration, regardless of when they were terminated.

Change-Id: I431232a0237ed71a8d19b53770b8ff21983a5a10