[GCE] Implement RPC service
Replaced Create with Ensure. The latter will be simpler as an internal
API for ensuring the datastore matches the imported config.

Generalized VMs entity to allow arbitrary IDs. The service will use
prefix after it validates prefix uniqueness when importing the config,
but this doesn't matter from the entity's perspective.

Bug: 897355
Change-Id: Idd7618127acd20b8a19d3a4fd11e7f184d29af91