Update patch set 1
Patch Set 1: Code-Review+1 Commit-Queue+2

> Patch Set 1: Auto-Submit+1
> 
> originally we didn't want to implement this because we thought it would be a lost investment: implementing the callback at the butler level would be more general, and reused in Robbie's work. However this is 4 LOC. I am not worried about this "loss of investment" as it enables progress of other important tasks

It was more like I'm planning on moving annotee out of kitchen in a couple weeks, so this will only be helpful for that and we also need a long term solution; it's fine as a small hack for now (since annotee is a hack).

Why []byte and not proto.Message?

Patch-set: 1
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1
Label: Commit-Queue=+2