[annotee] Implement callback
Implement a callback that is called in the same code path which sends the
serialized annotation proto to butler as a datagram.

Keep it seralized in the callback in case the callback does not need to
deserialize it either.

Bug: 843714
Change-Id: Ib62e43b3e81763c4c36d8834e101668bae5b66e8
Reviewed-on: https://chromium-review.googlesource.com/c/1328743
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>