[cipd] Calculate instance ID while building the package.
This is part 1 of some API simplification around RegisterInstance call.
It currently requires using deployer.OpenPackage during uploads (to get
pkg.Instance), which is an overkill. All we need is a pin, we don't need
to read zip directory just to upload a package.

R=nodir@chromium.org, iannucci@chromium.org, tandrii@chromium.org

Change-Id: Idf94fd7951f4b717937b0ee40570d293bce6b8fc
Reviewed-on: https://chromium-review.googlesource.com/c/1334209
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>