[cipd] Extract OpenInstance and friends from 'deployer' into 'reader'.
No new code, just code moves.

There are cases when we want to read the package without necessarily deploying
it. In particular, there's a scenario in which we want to read the package
manifest before uploading the package. Importing 'deployer' module for this
looks weird.

R=iannucci@chromium.org, nodir@chromium.org, tandrii@chromium.org

Change-Id: Ia700f9c8dd1180f4191e1ba1f656a33688da3081
Reviewed-on: https://chromium-review.googlesource.com/c/1338520
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>