[prpc] Enable CORB protection for pRPC responses
https://chromium.googlesource.com/chromium/src/+/master/services/network/cross_origin_read_blocking_explainer.md
recommends JSON responses to use a valid JSON media types in Content-Type header
and include "X-Content-Type-Options: nosniff" header.
Together they mitigate spectre attack.

Add pRPC protocol v1.1 which defines that.
This protocol change is backward compatible, so we can just patch the server
code without introducing a formal version header.
The patching will be done in a separate CL.

Note that from the CORB point of view, the media type used in **request**
does not matter, so this proposal is limited to responses.
Prohibiting "application/prpc; encoding=json" altogether would be backward incompatible,
which would defeat the benefits of this simpler design, compared to go/prpc-2.

Bug: 853323
Change-Id: I42a3827cad2720e99e0094dfdac388218c1d52aa