Update patch set 1
Patch Set 1:

Any thoughts on a good way to test this? Ideally we'd construct the Butler with a Butler Config with a StreamRegistrationCallback, and then check that the Butler's Bundler's Bundler Config had the right StreamChunkCallback, but the latter is not publicly accessible (specifically, a Bundler's Config is internal to the Bundler).

Alternatively we could test that wrapStreamRegistrationCallback follows the right branches, but that's not exactly an enlightening test.

Patch-set: 1
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false