[starlark] Refactor Interpreter to be more structured.
This introduces a notion of a package, which is a bunch of starlark modules
under a single root that can import each other and which are distributed and
versioned as a single unit.

This is roughly similar to how Bazel structures *.bzl code as far as my research
shows.

There are two special packages for now:
  "__main__": an alias for top-level user supplied code, since all code have
  to live in some package.
  "stdlib": a starlark code that implements various global functions, it is
  auto-executed by the interpreter before doing anything else.

Presumably in the (far) future we'll be able to do something like:

  external_cipd_package(alias="stuff", pkg="infra...", version="latest")
  load("@stuff//some/starlark.star", "say_hi")
  say_hi()

(Again, roughly similar to Bazel's http_archive).

Also protobuf handling is now less special. All protobuf messages embedded in
the binary are now represented by a starlark package (conventionally named
"proto"):

  load("@proto//go.chromium.org/luci/.../config.proto", 'buildbucket')
  buildbucket.Bucket(...)

The interpreter itself is oblivious of proto stuff.

R=iannucci@chromium.org
BUG=833946

Change-Id: Ia0608a9ab26fdf025b61a65377722d587ee3fc80
Reviewed-on: https://chromium-review.googlesource.com/c/1344645
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>