Update patch set 4
Patch Set 4:

Vadim, nice to see that you started working on this. Are you still experimenting or are we at the stage where we make longer-term design decisions? From the CL description, it seems it is the latter? Is there an updated design doc? I'd like to be able to follow/review the design/api-level decisions without going into code/impl-level details. How can I do that?

Patch-set: 4
CC: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>