Update patch set 4
Patch Set 4:

One more thing, on a wide monitor, there is a lot of wasted horizontal space due to some of the stuff hiding behind tabs.  I'm certain I will not be the last person to mention this (wasted horizontal space was some of the top feedback we received on an early version of the console view)
https://screenshot.googleplex.com/aNV2dHwM8mD

My idea would be to hide or show properties / changes on the same tab as the result depending on the width of the viewport.
Eg for narrow windows, 4 tabs is fine
for wide windows, just do result + timeline, where properties/changes are in results

Patch-set: 4
Reviewer: Gerrit User 1002080 <1002080@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>