[starlark] Allow propagating context.Context to builtins.
In lucicfg, don't bind *State to builtins, but pass it through the context. This
allows to load starlark code only once, but execute it multiple times
manipulating different state.

We don't really need it now, but this is cleaner and appears to be a use case
intended by Starlark go library.

R=iannucci@chromium.org, nodir@chromium.org
BUG=833946

Change-Id: I28cb8ead5513abd3654052610d72df07a48fc5c0
Reviewed-on: https://chromium-review.googlesource.com/c/1347489
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>