[lucicfg] Low-level API for generating arbitrary configs.
This is an escape hatch for cases when a high-level declarative API is not
sufficient. It will also be used internally by the high-level generator itself.

Config generating functions are registered as callbacks to allow high-level
configs to be defined in a declarative way, where order of definitions is not
important, as long as the entire high-level config is consistent. To allow this,
we should execute all top-level Starlark code before doing any evaluation of
the state, thus a necessity for delaying generator calls via callbacks.

R=iannucci@chromium.org, nodir@chromium.org
BUG=833946

Change-Id: Ib62640a316fc1255de873defc521be7cf67665bd