[starlark] Add 'stacktrace(skip=0)' builtin.
It captures a stack trace at the moment it was called and returns it as a value,
so it can be rendered at some later time.

Will be used to "remember" where various lucicfg rules were defined, so that
at a later time we can emit useful error messages.

R=iannucci@chromium.org
BUG=833946

Change-Id: I52d6cf694fc4a454c1534695a8f9ecc53e154c5c
Reviewed-on: https://chromium-review.googlesource.com/c/1354521
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>