Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > I see what's going on, so the tab call is actually happening first, but for some reason the page is refusing to render anything until after the time.js code finished running:
> > 
> > https://screenshot.googleplex.com/88AnkZWDJDp
> > 
> > This shouldn't be necessary, the page can render with tabs, but without time.js.  That should bring down the perceived latency.  Any idea what we can do here?
> 
> This script delivers the code that enables the milo.makeTimesLocal() and milo.annotateDurations() calls towards the bottom of the base.html page correct?
> 
> There are probably 3 obvious options to try:
> 
> 1) move the tag even lower, to the bottom or even outside the body tag.
> 2) use defer on the script tag
> 3) use async on the script tag
> 
> of course these need to work with the jquery document ready callback that uses that code. I will look at these approaches later today.

My suggestions above only work to change the time when time.js loads, not that useful. Those time functions still need to fire and that is where the slow down occurs.

I have a fix in place that puts the time manipulation functions in a callback so that the browser will finish the tab rendering quickly and then do the time fix up momentarily. This makes the delay from empty page rendered to tabs rendered under 100 ms. PTAL:

https://3627-e4e2ba7-tainted-dburger-dot-luci-milo-dev.appspot.com/p/fuchsia/builders/luci.fuchsia.ci/garnet-roller/b8941456073601677008

Patch-set: 9