[cipd] Handle in-place upgrade of files into directories.
Consider a package instance v1 that has only one file "a/b/c", and a package
instance v2 that has only one file "a/b/c/d". Before this CL, when upgrading
v1 to v2, CIPD would first create "a/b/c/d" as a new file (converting "a/b/c" to
a directory as a side effect), and then attempt to remove "a/b/c" (since it's
no longer in the package), and hit an error.

This situation is now recognized properly and CIPD doesn't try to remove such
directories.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=911229

Change-Id: I86b73a2746c2af89cc6b39644d25b629697fa8a8
Reviewed-on: https://chromium-review.googlesource.com/c/1359722
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>