[luci_notify] Read email_notify from properties
In the past we've decided to pass email_notify list via parameters, as opposed
to properties. V2 does not have parameters, only properties, and there are no
other use cases to extend go/build-proto.

It is possible that a recipe may want to mutate the email_notify list, e.g.
define who to notify, thus it makes sense to have the email notify list in
build output too.

To avoid polluting go/build-proto for a luci-notify-specific feature, keep
it a "schemaless" property, as opposed to a proto field in go/build-proto.

Parse email_notify from output. If it isn't there, read it from input.
Otherwise fallback to parameters.
ScheduleBuild v2 API does not allow passing parameters, so eventually users
will be forced to switch to properties.

Bug: 875056
Change-Id: I36da0eb11743f368424132ccb5c8bda37d0ffde6